refactor 统一返回类型

+++ src/main/java/com/daniu/controller/BlogController.java
@@ -2,7 +2,8 @@ package com.daniu.controller;
 
 import com.daniu.common.constant.Constants;
 import com.daniu.common.exception.BusinessException;
-import com.daniu.common.response.Result;
+import com.daniu.common.response.R;
+import com.daniu.domain.entity.MapData;
 import com.daniu.service.MapDataService;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -36,12 +37,12 @@ public class BlogController {
      * 查询单个文件
      *
      * @param name 文件名，不包含路径，文件类型应为JSON。
-     * @return {@link Result }
+     * @return {@link R }<{@link JsonNode }>
      * @throws IOException IOException
      */
     @GetMapping("selectOne")
     @Operation(summary = "根据id查询")
-    public Result selectOne(String name) throws IOException {
+    public R<JsonNode> selectOne(String name) throws IOException {
         // 构造文件完整路径
         String filePath = Constants.DOC_PATH + name;
         // 读取文件全部内容
@@ -50,22 +51,22 @@ public class BlogController {
         // 将内容解析为JSON节点
         JsonNode jsonNode = objectMapper.readTree(content);
 
-        return Result.success("查询成功", jsonNode);
+        return R.success("查询成功", jsonNode);
     }
 
     /**
      * 上传文件
      *
      * @param file 用户上传的文件，必须是 .md 格式。
-     * @return {@link Result }
-     * @apiNote 上传单个markdown文件到服务器，并处理该文件。
+     * @return {@link R }<{@link String }>
+     * @throws IOException IOException
      */
     @PostMapping("/insertOne")
     @Operation(summary = "上传markdown文件")
-    public Result insertOne(@RequestParam("file") MultipartFile file) throws IOException {
+    public R<String> insertOne(@RequestParam("file") MultipartFile file) throws IOException {
         if (file.isEmpty()) throw new BusinessException("请上传文件");
         mapDataService.insertOrUpdateMapData(file);
-        return Result.success("文件上传成功");
+        return R.ok("文件上传成功");
     }
 
 
@@ -73,20 +74,24 @@ public class BlogController {
      * 删除文件
      *
      * @param id   id
-     * @param name 文件名
-     * @return {@link Result }
+     * @return {@link R }<{@link String }>
      * @throws IOException IOException
      */
     @DeleteMapping("/deleteOne")
     @Operation(summary = "删除markdown文件")
-    public Result deleteOne(Integer id, String name) throws IOException {
-        String filePath = Constants.DOC_PATH + name;
+    public R<String> deleteOne(Integer id) throws IOException {
+        MapData mapData = mapDataService.getById(id);
+        if (mapData == null) throw new BusinessException("文件不存在");
+
+        String jsonFilePath = Constants.DOC_PATH + mapData.getPath();
+        String filePath = Constants.DOC_PATH + mapData.getName();
 
         boolean removed = mapDataService.removeById(id);
         if (!removed) throw new BusinessException("删除失败");
 
+        Files.deleteIfExists(Paths.get(jsonFilePath));
         Files.deleteIfExists(Paths.get(filePath));
-        return Result.success("删除成功");
+        return R.ok("删除成功");
     }
 
 }
